feat: combines create function with control
